requirements.txt
colorama==0.4.3
certifi==2020.4.5.2
chardet==3.0.4
dlib==19.20.0
future==0.18.2
idna==2.9
numpy==1.18.5
opencv-python==4.1.2.30
PyYAML==5.1.2
regex==2020.6.8
requests==2.23.0
rope==0.16.0
torch==1.5.0
urllib3==1.25.9
pandas==0.25.2
youtube-dl==2020.1.24
scipy==1.4.1
scikit-learn==0.22.1
matplotlib==3.1.1
moviepy==1.0.0
tensorflow==2.1.0
python-include==1.0.0